INKO-329 (#537)
* INKO-329

+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/collateralmanagement/service/ClientCollateralManagementWritePlatformServiceImpl.java
@@ -291,11 +291,11 @@ public class ClientCollateralManagementWritePlatformServiceImpl implements Clien
         final ClientCollateralManagement clientCollateralManagement = this.clientCollateralManagementRepositoryWrapper
                 .getCollateral(collateralId);
         validateForDeletion(clientCollateralManagement, collateralId);
-        ClientCollateralManagementAdditionalDetails details = this.clientCollateralManagementAdditionalDetailsRepository
-                .findByCollateralId(clientCollateralManagement);
-        this.clientCollateralManagementAdditionalDetailsRepository.delete(details);
-        ;
 
+        ClientCollateralManagementAdditionalDetails details = this.clientCollateralManagementAdditionalDetailsRepository.findByCollateralId(clientCollateralManagement);
+        if (details != null) {
+            this.clientCollateralManagementAdditionalDetailsRepository.delete(details);;
+        }
         this.clientCollateralManagementRepositoryWrapper.deleteClientCollateralProduct(collateralId);
         return new CommandProcessingResultBuilder().withEntityId(collateralId).build();
     }
